SigmaEstablisher works with Advertisement and Messages on choix (asynchrone or synchrone) 	modifié :         src/main/java/network/api/EstablisherService.java 	modifié :         src/main/java/network/api/EstablisherServiceListener.java 	modifié :         src/main/java/network/impl/jxta/JxtaEstablisherService.java 	renommé :         src/main/java/protocol/impl/SigmaEstablisherAsync.java -> src/main/java/protocol/impl/SigmaEstablisher.java 	modifié :         src/main/java/protocol/impl/sigma/Trent.java 	supprimé :        src/main/java/protocol/impl/sigma/TrentAsync.java 	modifié :         src/main/java/protocol/impl/sigma/TrentSolver.java 	supprimé :        src/test/java/controller/ApplicationForTests.java 	renommé :         src/test/java/protocol/impl/SigmaEstablisherAsyncFailer.java -> src/test/java/protocol/impl/SigmaEstablisherFailer.java 	renommé :         src/test/java/protocol/impl/SigmaEstablisherAsyncTest.java -> src/test/java/protocol/impl/SigmaEstablisherTest.java
