+++ src/main/java/network/api/EstablisherService.java
@@ -3,6 +3,9 @@
  */
 package network.api;
 
+import java.util.HashMap;
+
+import model.entity.ElGamalKey;
 import network.api.service.Service;
 import network.impl.messages.EstablisherMessage;
 
@@ -14,6 +17,17 @@ public interface EstablisherService extends Service
 {
 	
 	public static final String NAME = "establisher";
+	
+	
+	/**
+	 * Encapsulate contract sending (using either Message or Advertisement)
+	 * @param title : event title
+	 * @param data : data sent
+	 * @param senderK : sender id 
+	 * @param uris : receiving peers uris, if null : uses Advertisement, otherwise uses Messages
+	 * @param peer : the sending peer
+	 */
+	public void sendContract(String title, String data, String senderK, HashMap<ElGamalKey, String> uris, Peer peer);
 	
 	/**
 	 * Send a "Message" (synchrone)
@@ -41,6 +55,16 @@ public interface EstablisherService extends Service
 	 */
 	public void sendContract(String title, String data, String sourceKey, Peer peer);
 
+	
+	
+	/**
+	 * Encapsulate listening system
+	 * @param field : field to be matched with @param value
+	 * @param listenerId : an id to listen on and to be able to remove the listener
+	 * @param l : listener notified upon receiving data
+	 * @param useMessage : if true -> use Messages, else use Advertisement
+	 */
+	public void setListener(final String field, final String value, String listenerId, final EstablisherServiceListener l, boolean useMessage);
 	/**
 	 * Add an advertisement listener and search in already sent adverts
 	 * @param field : field to be matched with @param value
@@ -48,9 +72,10 @@ public interface EstablisherService extends Service
 	 * @param l : listener notified when advert received 
 	 */
 	public void listens(String field, String value, String listenerId, EstablisherServiceListener l);
+	
 	/*
 	 * Remove the previously set listener
 	 */
-	public void removeListens(String listenerId);
+	public void removeListener(String listenerId);
 	
 }
