+++ src/main/java/protocol/impl/SigmaEstablisher.java
@@ -8,6 +8,8 @@ import java.util.HashMap;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 
+import crypt.impl.signatures.ElGamalSignature;
+import crypt.impl.signatures.ElGamalSigner;
 /* TODO : Respect design pattern factory here
  * 		Problem here is that the implementation use other keys than the main one
  * 		Same problem in Trent
@@ -19,13 +21,14 @@ import controller.Application;
 import controller.tools.JsonTools;
 import crypt.api.encryption.Encrypter;
 import crypt.factories.EncrypterFactory;
+import crypt.factories.SignerFactory;
 import model.api.Status;
 import model.entity.ContractEntity;
 import model.entity.ElGamalKey;
 import model.entity.sigma.Or;
 import network.api.EstablisherService;
-import network.api.Messages;
-import network.api.ServiceListener;
+import network.api.EstablisherServiceListener;
+import network.api.Peer;
 import protocol.api.Establisher;
 import protocol.impl.sigma.PCS;
 import protocol.impl.sigma.Sender;
@@ -42,12 +45,19 @@ import protocol.impl.sigma.SigmaContract;
 
 public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaSignature, SigmaSigner, SigmaContract> {
 	
+	public static final String STARTING_MESSAGE = "START";
+	public static final String TRENT_MESSAGE = "TRENT";
+	public static final String FOR_TRENT_MESSAGE = "FOR TRENT";
+	public static final String SIGNING_MESSAGE = "SIGNING_DATA";
 	
 	protected ElGamalKey trentK;
-	protected final EstablisherService establisherService =(EstablisherService) Application.getInstance().getPeer().getService(EstablisherService.NAME);
+	protected EstablisherService establisherService =(EstablisherService) Application.getInstance().getPeer().getService(EstablisherService.NAME);
+	protected Peer peer = Application.getInstance().getPeer();
+
 	
 	// Store the different rounds of the signing protocol
 	protected Or[][] promRoundSender;
+	// Choose whether to use message or advertisement
 	
 	// Id for the contract (make sure we sign the same)
 	protected String contractId;
@@ -60,10 +70,9 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 	// Know which parties are ready to sign
 	protected String[] ready;
 	// Current signature round
-	protected int round = 0;
+	protected int round = -1;
 	// The PCS maker
 	protected PCS pcs; 
-	
 	
 	
 	/**
@@ -72,11 +81,13 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 	 * @param <uri> : parties matching uri
 	 */
 	// TODO : REMOVE trentK
-	public SigmaEstablisher(ElGamalKey senderK, HashMap<ElGamalKey, String> uri, ElGamalKey t){
-		signer = new SigmaSigner();
-		signer.setKey(senderK);
-		uris = uri;
-		trentK = t;
+	public SigmaEstablisher(ElGamalKey senderK, ElGamalKey t, HashMap<ElGamalKey,String> uris){
+		this.signer = SignerFactory.createSigmaSigner();
+		this.signer.setKey(senderK);
+		
+		this.uris = uris;
+		
+		this.trentK = t;
 	}
 	
 	/**
@@ -93,6 +104,7 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 		// The El Gamal keys of all parties 
 		keys = contract.getParties();
 		senPubK = signer.getKey().getPublicKey();
+		
 		N = keys.size();
 		// Used to know wether everyone has yet accepted the contract or not
 		ready = new String[N];
@@ -101,26 +113,35 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 		
 		/*
 		 * Get ready to start
-		 * If we receive a starting message, it is stored and we wait until everyone has sent
-		 * 		its starting message
+		 * If an advertisement was or is received, we check the signature and it is stored and we wait until everyone has sent its starter
 		 */
-		establisherService.addListener(new ServiceListener() {
+		establisherService.removeListener(STARTING_MESSAGE+contractId+senPubK.toString());
+		establisherService.setListener("title", STARTING_MESSAGE+contractId, STARTING_MESSAGE+contractId+senPubK.toString(), new EstablisherServiceListener(){
 			@Override
-			public void notify(Messages messages) {// Finding the sender
-				BigInteger msgSenKey = new BigInteger(messages.getMessage("sourceId"));
+			public void notify(String title, String data, String senderId) {
+				JsonTools<ElGamalSignature> json = new JsonTools<>(new TypeReference<ElGamalSignature>(){});
+				ElGamalSignature signature = json.toEntity(data);
+				
+				BigInteger msgSenKey = new BigInteger(senderId);
+	
 				int i = 0;
 				while (!(keys.get(i).getPublicKey().equals(msgSenKey))){i++;}
 				
-				if (decryptMsg(messages.getMessage("contract"),signer.getKey()).equals("start") &&
-					messages.getMessage("title").equals(contractId)){
+				// Prepare the elGamalSigner to check the data
+				ElGamalSigner elGamalSigner = SignerFactory.createElGamalSigner();
+				elGamalSigner.setKey(keys.get(i));
+
+				if (elGamalSigner.verify(STARTING_MESSAGE.getBytes(), signature)){
 					ready[i] = "";
 					// Checks if everyone is ready
-					if (Arrays.asList(ready).indexOf(null) == (-1)){
+					if (Arrays.asList(ready).indexOf(null) == (-1) && round==-1){
+						round++;
 						chooseTrent();
 					}
+				
 				}
 			}
-		}, senPubK.toString());
+		}, uris != null);
 	}
 	
 	/**
@@ -128,13 +149,21 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 	 */
 	@Override
 	public void start(){
-		for (int k=0; k<N; k++){
-			ElGamalKey key = keys.get(k);
-			establisherService.sendContract(contractId,
-											key.getPublicKey().toString(), 
-											senPubK.toString(),
-											encryptMsg("start",key),
-											uris.get(key));
+		JsonTools<ElGamalSignature> json = new JsonTools<>(new TypeReference<ElGamalSignature>(){});
+		
+		ElGamalSigner elGamalSigner = SignerFactory.createElGamalSigner();
+		elGamalSigner.setKey(signer.getKey());
+		// Creates signature on message "start" and change it to Json
+		String content = json.toJson(elGamalSigner.sign((STARTING_MESSAGE).getBytes()));
+
+		// Sending an advertisement
+		establisherService.sendContract(STARTING_MESSAGE+contractId, content, senPubK.toString(), uris, peer);
+		int i=0;
+		while (!(keys.get(i).getPublicKey().equals(senPubK))){i++;}
+		ready[i] = "";
+		if (Arrays.asList(ready).indexOf(null) == (-1) && round == -1){
+			round++;
+			chooseTrent();
 		}
 	}
 
@@ -142,33 +171,37 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 	 * Choose Trent, put a listener for him, then start signing
 	 */
 	protected void chooseTrent(){
+		establisherService.removeListener(STARTING_MESSAGE+contractId+senPubK.toString());
+		
 		// TODO : choose trent and associate its uri in uris
 		contract.setTrentKey(trentK);
 		signer.setTrentK(trentK);
 		
 		
 		// Put a listener on Trent in case something goes wrong
-		establisherService.addListener(new ServiceListener(){
+		establisherService.removeListener(TRENT_MESSAGE+contractId+senPubK.toString());
+		establisherService.setListener("title", TRENT_MESSAGE + contractId, TRENT_MESSAGE+contractId+senPubK.toString(), new EstablisherServiceListener(){
 			@Override
-			public void notify(Messages messages){
-				
-				// If the message is for another contract
-				if (messages.getMessage("title").equals(contractId)){
-					
+			public void notify(String title, String data, String senderId) {
+				// If the message is for another contract or by someone else thant Trent
+				if (senderId.equals(trentK.getPublicKey().toString())){
 					// If Trent found we were dishonest (second time a resolve sent)
-					if (messages.getMessage("contract").equals("Dishonest")){
-						System.out.println("You were dishonest, third party didn't do nothing");
+					if (data.equals("Dishonest")){
+						System.out.println("You were dishonest or request sent twice, third party didn't do nothing on this time");
 					} 
 					
 					else{
 						JsonTools<ArrayList<String>> jsons = new JsonTools<>(new TypeReference<ArrayList<String>>(){});
-						ArrayList<String> answer = jsons.toEntity(messages.getMessage("contract"));
+						ArrayList<String> answer = jsons.toEntity(data);
 
 						// Making sure the message is from Trent
 						signer.setTrentK(trentK);
 						JsonTools<SigmaSignature> json = new JsonTools<>(new TypeReference<SigmaSignature>(){});
 
-						if(signer.verify(answer.get(1).getBytes() ,json.toEntity(answer.get(2)))){
+						JsonTools<HashMap<String,String>> jsonH = new JsonTools<>(new TypeReference<HashMap<String,String>>(){});
+						HashMap<String,String> signatures = jsonH.toEntity(answer.get(2));
+						
+						if(signer.verify(answer.get(1).getBytes() ,json.toEntity(signatures.get(senPubK.toString())))){
 							// If Trent aborted the contract
 							if (answer.get(0).equals("aborted") || answer.get(0).equals("honestyToken")){
 								setStatus(Status.CANCELLED);
@@ -182,13 +215,13 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 
 								
 								// Check the signatures (we don't if it was on round -1 or -2)
-								byte[] data = (new String(contract.getHashableData()) + (round - 1)).getBytes();
+								byte[] data1 = (new String(contract.getHashableData()) + (round - 1)).getBytes();
 								byte[] data2 = (new String(contract.getHashableData()) + (round - 2)).getBytes();
 								
 								for (SigmaSignature signature : sigSign){
 									signer.setReceiverK(keys.get(sigSign.indexOf(signature)));
 									
-									if (signer.verify(data, signature) || signer.verify(data2, signature)){
+									if (signer.verify(data1, signature) || signer.verify(data2, signature)){
 										contract.addSignature(keys.get(sigSign.indexOf(signature)), signature);
 									}
 								}
@@ -203,7 +236,7 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 					}
 				}
 			}
-		}, senPubK + "TRENT");
+		}, false);
 		
 		sign();
 	}
@@ -220,35 +253,32 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 		pcs = new PCS(sender, trentK);
 		round = 1;
 
-		establisherService.addListener(new ServiceListener() {
+		establisherService.removeListener(SIGNING_MESSAGE+contractId+senPubK.toString());
+		establisherService.setListener("title", SIGNING_MESSAGE+contractId, SIGNING_MESSAGE+contractId+senPubK.toString(), new EstablisherServiceListener() {
 			@Override
-			public void notify(Messages messages) {
-				if (messages.getMessage("title").equals(contractId)){
-					String msg = messages.getMessage("contract");
-					
-					// Checks if the message is a PCS, if yes store it in "pcs[round][k]"
-					verifyAndStoreSignature(msg, messages.getMessage("sourceId"));
+			public void notify(String title, String msg, String senderId) {
+				// Checks if the message is a PCS, if yes store it in "pcs[round][k]"
+				verifyAndStoreSignature(msg, senderId);
 
-					// Check if the round is complete
-					boolean claimFormed = true;
+				// Check if the round is complete
+				boolean claimFormed = true;
+				for (int k=0; k<N; k++){
+					if (promRoundSender[round][k] == null)
+						claimFormed= false;
+				}
+				
+				/*	Send the rounds (if we have the claim needed):
+				 *  	We do a loop because sometimes, we receive the PCS for round+1 before the one for the current round
+				 */  
+				while (round<=(N+1) && claimFormed){
+					sendRound(++round);
 					for (int k=0; k<N; k++){
 						if (promRoundSender[round][k] == null)
 							claimFormed= false;
 					}
-					
-					/*	Send the rounds (if we have the claim needed):
-					 *  	We do a loop because sometimes, we receive the PCS for round+1 before the one for the current round
-					 */  
-					while (round<=(N+1) && claimFormed){
-						sendRound(++round);
-						for (int k=0; k<N; k++){
-							if (promRoundSender[round][k] == null)
-								claimFormed= false;
-						}
-					}
 				}
 			}
-		}, senPubK.toString());
+		}, uris != null);
 
 		// Send the first round
 		sendRound(1);
@@ -266,52 +296,41 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 	 * Trent resolve function is in Trent Class
 	 */
 	protected void resolve(){
-
-		establisherService.removeListener(senPubK.toString());
+		establisherService.removeListener(SIGNING_MESSAGE+contractId+senPubK.toString());
 		
-		String[] content = new String[5];
+		String[] content = new String[4];
 
 		// Round
 		content[0] = String.valueOf(round-1);
-		// Uris
-		JsonTools<HashMap<BigInteger, String>> json1 = new JsonTools<>(new TypeReference<HashMap<BigInteger, String>>(){});
-		HashMap<BigInteger, String> u = new HashMap<BigInteger, String>();
-		for (ElGamalKey ke : uris.keySet()){
-			u.put(ke.getPublicKey(), uris.get(ke));
-		}
-		content[1] = json1.toJson(u);
 		
 		
 		// Contract
 		JsonTools<ContractEntity> json2 = new JsonTools<>(new TypeReference<ContractEntity>(){});
-		content[2] = json2.toJson(contract.getEntity(),false);
+		content[1] = json2.toJson(contract.getEntity(),false);
 		
 		// Claim(k)
 		signer.setReceiverK(trentK);
 		SigmaSignature sigClaimK;
 		if (round<=1){
-			content[3] = encryptMsg("ABORT", trentK);
+			content[2] = encryptMsg("ABORT", trentK);
 			sigClaimK = signer.sign("ABORT".getBytes());
 		}else {
 			JsonTools<Or[]> json = new JsonTools<>(new TypeReference<Or[]>(){});
 			String claimK = json.toJson(promRoundSender[round-1], true);
-			content[3] = encryptMsg(claimK, trentK);
+			content[2] = encryptMsg(claimK, trentK);
 			sigClaimK = signer.sign(claimK.getBytes());
 		}
 		JsonTools<SigmaSignature> json3 = new JsonTools<>(new TypeReference<SigmaSignature>(){});
-		content[4] = encryptMsg(json3.toJson(sigClaimK, false), trentK);
+		content[3] = encryptMsg(json3.toJson(sigClaimK, false), trentK);
 		
 		// Concatenate the content
 		JsonTools<String[]> json = new JsonTools<>(new TypeReference<String[]>(){});
 		String fullContent = json.toJson(content, false);
 
 		System.out.println("--- Sending resolve request to Trent --- Round : " + (round-1));
-		
-		establisherService.sendContract(contractId,
-							trentK.getPublicKey().toString()+"TRENT",
-							senPubK.toString(),
-							fullContent,
-							uris.get(trentK));
+
+		// For Trent, use only Advertisement
+		establisherService.sendContract(FOR_TRENT_MESSAGE + trentK.getPublicKey().toString(), fullContent, senPubK.toString(), null, peer);
 	}
 	
 	/*
@@ -321,6 +340,10 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 	 */
 	protected void sendRound(int round){
 		// Loop : send to every party the correct message
+
+		// Content of the message which will be sent
+		HashMap<String, String> content = new HashMap<String, String>();
+		
 		for (int k=0; k<N; k++){
 
 			// Public key of the receiver
@@ -328,9 +351,6 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 			
 			// If the receiver is the sender, isSender = true 
 			boolean isSender = receiverK.getPublicKey().equals(senPubK);
-
-			// Content of the message which will be sent
-			String content;
 			
 			// On the last round, send the clear signature
 			if (round==(N+2)){
@@ -338,36 +358,32 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 				SigmaSignature signature = pcs.getClearSignature(contract, receiverK);
 				if (isSender)
 					contract.addSignature(keys.get(k), signature);
-				content = json.toJson(signature, true);
+				content.put(receiverK.getPublicKey().toString(), json.toJson(signature, true));
 			// Otherwise send round k
 			}else {
 				byte[] data = (new String(contract.getHashableData()) + round).getBytes();
 				Or p = pcs.getPcs(data, receiverK, true);
-				content=encryptMsg(getJson(p), receiverK);
+				content.put(receiverK.getPublicKey().toString(), encryptMsg(getJson(p), receiverK));
 				if (isSender){
 					promRoundSender[round][k] = p;
 				}
 			} 
-			
-			// No need to send if we are the receiver
-			if (!isSender){
-				// Encapsulate the data we shall send (round + content)
-				String[] fullContent = {String.valueOf(round), content};
-				JsonTools<String[]> json2 = new JsonTools<>(new TypeReference<String[]>(){});
-				String fullContentS = json2.toJson(fullContent);
-
-				// Getting the sender public key index 
-				int i = 0;
-				while (!(keys.get(i).getPublicKey().equals(senPubK))){i++;}
-				System.out.println("Sending Round : " + round + " - for " + k + " : by " + i);
-				
-				establisherService.sendContract(contractId, 
-									receiverK.getPublicKey().toString(),
-									senPubK.toString(),
-									fullContentS,
-									uris.get(receiverK));
-			}
 		}
+		
+		// Adding the round to data sent
+		content.put("ROUND", String.valueOf(round));
+		
+		// Convert map to String
+		JsonTools<HashMap<String,String>> json2 = new JsonTools<>(new TypeReference<HashMap<String,String>>(){});
+		String dataToBeSent = json2.toJson(content);
+		
+		// Getting the sender public key index 
+		int i = 0;
+		while (!(keys.get(i).getPublicKey().equals(senPubK))){i++;}
+		System.out.println("Sending Round : " + round + " : by " + i);
+		
+		// Sending an advertisement
+		establisherService.sendContract(SIGNING_MESSAGE + contractId, dataToBeSent, senPubK.toString(), uris, peer);
 	}
 	
 	/*
@@ -385,20 +401,19 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 		int i = 0;
 		while (!(keys.get(i).getPublicKey().equals(msgSenKey))){i++;}
 		ElGamalKey senderKey = keys.get(i);
-		
+
 		// From json message to the object {"k", PCS}
-		JsonTools<String[]> json0 = new JsonTools<>(new TypeReference<String[]>(){});
-		String[] fullContent = json0.toEntity(message, true);
+		JsonTools<HashMap<String,String>> json0 = new JsonTools<>(new TypeReference<HashMap<String,String>>(){});
+		HashMap<String,String> content = json0.toEntity(message);
 		
-		int k= Integer.parseInt(fullContent[0]);
-		String content = fullContent[1];
+		int k= Integer.parseInt(content.get("ROUND"));
 		
 		// Don't do anything if the sender is the actual user (shouldn't happen though)
 		if (!(senderKey.getPublicKey().equals(senPubK))){
 			// If it's the last round, test the clear signature
 			if (k == (N+2)){
-				JsonTools<SigmaSignature> json = new JsonTools<>(new TypeReference<SigmaSignature>(){});
-				SigmaSignature signature =  json.toEntity(content, true);
+				JsonTools<SigmaSignature> json2 = new JsonTools<>(new TypeReference<SigmaSignature>(){});
+				SigmaSignature signature =  json2.toEntity(content.get(senPubK.toString()), true);
 				
 				SigmaSigner s = new SigmaSigner();
 				ElGamalKey ke = new ElGamalKey();
@@ -411,13 +426,13 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 					if (contract.isFinalized()){
 						int j = 0;
 						while (!(keys.get(j).getPublicKey().equals(senPubK))){j++;}
-						System.out.println("--- CONTRACT FINALIZED -- id : " + j);
+						System.out.println("--- CONTRACT FINALIZED -- on round : "+ round + "  -- id : " + j);
 						setStatus(Status.FINALIZED);
 					}
 				}
 			// Otherwise, test if it is the correct PCS, if so : store it
 			}else {
-				String msg = decryptMsg(content, signer.getKey());
+				String msg = decryptMsg(content.get(senPubK.toString()), signer.getKey());
 				byte[] data = (new String(contract.getHashableData()) + k).getBytes();
 				if (getPrivateCS(msg).Verifies(data)){
 					promRoundSender[k][i]=getPrivateCS(msg);
@@ -431,7 +446,8 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 	 * Primitives
 	 * 		Transformation json <-> object
 	 * 		Decrypting <-> Encrypting
-	 */	
+	 */
+	
 	// Return the string representing the private contract signature
 	protected String getJson(Or pcs){
 		JsonTools<Or> json = new JsonTools<>(new TypeReference<Or>(){});
@@ -442,6 +458,7 @@ public class SigmaEstablisher extends Establisher<BigInteger, ElGamalKey, SigmaS
 		JsonTools<Or> json = new JsonTools<>(new TypeReference<Or>(){});
 		return json.toEntity(pcs, true);
 	}
+	
 	// Return the message encrypted with public key
 	protected String encryptMsg(String msg, ElGamalKey key){
 		Encrypter<ElGamalKey> encrypter = EncrypterFactory.createElGamalSerpentEncrypter();
