--- src/test/java/controller/ApplicationForTests.java
@@ -1,88 +1,0 @@
-package controller;
-
-import java.util.Properties;
-
-import controller.tools.LoggerUtilities;
-import model.syncManager.UserSyncManagerImpl;
-import network.api.Peer;
-import network.factories.PeerFactory;
-import rest.api.Authentifier;
-import rest.factories.AuthentifierFactory;
-import rest.factories.RestServerFactory;
-
-/**
- * Main class
- * {@link Application} is a singleton
- * @author Julien Prudhomme
- *
- */
-public class ApplicationForTests {
-	public final static int jxtaPort = 9801;
-	public final static int restPort = 8081;
-	public final static String[] rdvPeerIds = {"tcp://176.132.64.68:9800", "tcp://localhost:9800", "tcp://localhost:9801"};
-	
-	private static ApplicationForTests instance = null;
-	private static UserSyncManagerImpl umg;
-	private Peer peer;
-	private Authentifier auth;
-
-	
-	
-	public ApplicationForTests() {
-		if(instance != null) {
-			throw new RuntimeException("Application can be instanciate only once !");
-		}
-		instance = this;
-	}
-
-	public static ApplicationForTests getInstance()	{
-		return instance;
-	}
-
-	public void run() {
-		setPeer(PeerFactory.createDefaultAndStartPeer());
-		setAuth(AuthentifierFactory.createDefaultAuthentifier());
-		RestServerFactory.createAndStartDefaultRestServer(restPort); //start the rest api
-	}
-
-	public void runForTests(int restPort) {
-		Properties p = System.getProperties();
-		p.put("derby.system.home", "./.db-" + restPort + "/");
-		umg = new UserSyncManagerImpl(); //just init the db
-		umg.close();
-		umg = null;
-		try {
-			setPeer(PeerFactory.createDefaultAndStartPeerForTest(jxtaPort, rdvPeerIds));
-			setAuth(AuthentifierFactory.createDefaultAuthentifier());
-			RestServerFactory.createAndStartDefaultRestServer(restPort);
-		} catch (Exception e) {
-			LoggerUtilities.logStackTrace(e);
-		}		
-	}
-
-	public static void main(String[] args) {
-		new ApplicationForTests();
-		ApplicationForTests.getInstance().runForTests(restPort);
-	}
-	
-	public void stop(){
-		peer.stop();
-		instance = null;
-	}
-	
-	public Peer getPeer() {
-		return peer;
-	}
-
-	public void setPeer(Peer peer) {
-		this.peer = peer;
-	}
-
-	public Authentifier getAuth() {
-		return auth;
-	}
-
-	public void setAuth(Authentifier auth) {
-		this.auth = auth;
-	}
-}
